Crate polkadot_test_runtime

The Polkadot runtime. This can be compiled with #[no_std], ready for Wasm.

Modules

api

constants

Constant values used within the runtime.

pallet_test_notifier

The module that hosts all the FRAME types needed to add this pallet to a runtime.

xcm_config

Structs

AnyNetwork

BaseXcmWeight

BondingDuration

DisabledValidatorsThreshold

EndingPeriod

EpochDuration

ExistentialDeposit

ExpectedBlockTime

FirstMessageFactorPercent

GenesisConfig

IndexDeposit

LeasePeriod

MaxAuthorities

MaxInstructions

MaxLocks

MaxNominatorRewardedPerValidator

MaxReserves

MinVestedTransfer

MinimumPeriod

Offset

Origin

PalletInfo

Provides an implementation of PalletInfo to provide information about the pallet setup in the runtime.

Period

Prefix

RewardCurve

Runtime

RuntimeApi

RuntimeApiImpl

Implements all runtime apis for the client side.

SS58Prefix

SessionKeys

SessionsPerEra

SlashDeferDuration

SlotDuration

TransactionByteFee

UncleGenerations

Version

Enums

BalancesCall

Contains one variant per dispatchable that can be called by an extrinsic.

Call

Event

OriginCaller

ParasSudoWrapperCall

Contains one variant per dispatchable that can be called by an extrinsic.

StakerStatus

Indicates the initial status of the staker.

SudoCall

Contains one variant per dispatchable that can be called by an extrinsic.

TimestampCall

Contains one variant per dispatchable that can be called by an extrinsic.

Constants

BABE_GENESIS_EPOCH_CONFIG

The BABE epoch configuration at genesis.

VERSION

Runtime version (Test).

WASM_BINARY

WASM_BINARY_BLOATY

Traits

BuildStorage

Complex storage builder stuff.

GetLastTimestamp

Functions

native_version

Native version.

Type Definitions

Address

The address format for describing accounts.

AllModulesDeprecated

All modules included in the runtime as a nested tuple of types. Excludes the System pallet.

AllModulesWithSystemDeprecated

All modules included in the runtime as a nested tuple of types.

AllPallets

All pallets included in the runtime as a nested tuple of types. Excludes the System pallet.

AllPalletsWithSystem

All pallets included in the runtime as a nested tuple of types.

AuthorityDiscovery

AuthorityDiscoveryConfig

Authorship

Babe

BabeConfig

Balances

BalancesConfig

Block

Block type as expected by this runtime.

BlockId

BlockId type as expected by this runtime.

Claims

ClaimsConfig

Configuration

ConfigurationConfig

Dmp

Executive

Executive: handles dispatch to the various modules.

Extrinsic

Grandpa

GrandpaConfig

Hash

Header

Block header type as expected by this runtime.

Historical

Hrmp

Indices

IndicesConfig

Initializer

LocalOriginToLocation

Offences

ParaInclusion

ParaInherent

ParaSessionInfo

Paras

ParasDisputes

ParasShared

ParasSudoWrapper

Scheduler

Session

SessionConfig

SignedBlock

A Block signed with a Justification

SignedExtra

The SignedExtension to the basic transaction logic.

SignedPayload

The payload being signed in transactions.

Staking

StakingConfig

Sudo

SudoConfig

System

SystemConfig

TestNotifier

Timestamp

TransactionPayment

Ump

UncheckedExtrinsic

Unchecked extrinsic type as expected by this runtime.

Vesting

VestingConfig

Xcm